% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsontab.R
\name{from_json_or_tab}
\alias{from_json_or_tab}
\title{Load a file that may be in either JSON or tabular format}
\usage{
from_json_or_tab(txt, json = FALSE, tab = FALSE, header = FALSE,
  stringsAsFactors = TRUE, ...)
}
\arguments{
\item{txt}{a JSON string, a URL, or a JSON or tabular file}

\item{json}{indicates JSON input}

\item{tab}{indicates tabular input}

\item{...}{additional arguments to be passed to \code{fromJSON} or
\code{read.table}}
}
\value{
a list representing the input data
}
\description{
Regardless of the input format, a list will be returned
}
\seealso{
\link{\code{simplify_data_frames}}
}
