% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsontab.R
\name{simplify_data_frames}
\alias{simplify_data_frames}
\title{Simplify data frames in input data}
\usage{
simplify_data_frames(x)
}
\arguments{
\item{x}{a list of lists to be converted to a list of data frames.}
}
\value{
a list of data frames
}
\description{
Data loaded using \code{from_json_or_tab} will be parsed
as a list of lists. This function accepts a list generated by
\code{from_json_or_tab} and returns a list of data frames.
}
\seealso{
\link{\code{from_json_or_tab}}
}
